refactor(api): split agent chat handler helpers
Extract agent-chat render helpers, activity helpers, and SSE streaming helpers into focused files. Keep the HTTP handler file centered on request handling while preserving the existing API behavior.
