+++ internal/api/handler_agent_chat.go
@@ -14,8 +14,6 @@ import (
 
 	"github.com/hecate/agent-runtime/internal/agentadapters"
 	"github.com/hecate/agent-runtime/internal/agentchat"
-	"github.com/hecate/agent-runtime/internal/config"
-	"github.com/hecate/agent-runtime/internal/profiler"
 	"github.com/hecate/agent-runtime/internal/telemetry"
 )
 
@@ -349,10 +347,9 @@ func (h *Handler) HandleCreateAgentChatMessage(w http.ResponseWriter, r *http.Re
 		// Defensive: the constructor in NewHandler always sets
 		// agentChatRunner. This branch only fires for programmer error
 		// (e.g. a test handler built without it). The fallback runner
-		// has no approval recorder installed, so it preserves the
-		// legacy auto-approve behavior — same as before the approval
-		// recorder existed. Tighten in a follow-up if the fallback
-		// remains in use.
+		// has no approval coordinator installed and falls back to
+		// auto-approve. Tighten in a follow-up if the fallback remains
+		// in use.
 		runner = agentadapters.NewSessionManager()
 	}
 	result, runErr := runner.Run(runCtx, agentadapters.RunRequest{
@@ -542,265 +539,6 @@ func (h *Handler) HandleCreateAgentChatMessage(w http.ResponseWriter, r *http.Re
 	}
 	h.agentChatLive.publishSession(updated)
 	WriteJSON(w, http.StatusOK, AgentChatSessionResponse{Object: "agent_chat_session", Data: renderAgentChatSession(updated, h.agentChatSnapshotConfig())})
-}
-
-func renderAgentChatSessionSummary(session agentchat.Session) AgentChatSessionSummaryItem {
-	return AgentChatSessionSummaryItem{
-		ID:              session.ID,
-		Title:           session.Title,
-		AdapterID:       session.AdapterID,
-		DriverKind:      session.DriverKind,
-		NativeSessionID: session.NativeSessionID,
-		Workspace:       session.Workspace,
-		WorkspaceBranch: session.WorkspaceBranch,
-		Status:          session.Status,
-		MessageCount:    len(session.Messages),
-		CreatedAt:       formatOptionalTime(session.CreatedAt),
-		UpdatedAt:       formatOptionalTime(session.UpdatedAt),
-	}
-}
-
-type agentChatSnapshotConfig struct {
-	MaxTurnsPerSession int
-	MaxSessionDuration time.Duration
-	IdleTimeout        time.Duration
-}
-
-func agentChatSnapshotConfigFromServer(cfg config.ServerConfig) agentChatSnapshotConfig {
-	return agentChatSnapshotConfig{
-		MaxTurnsPerSession: cfg.AgentChatMaxTurnsPerSession,
-		MaxSessionDuration: cfg.AgentChatMaxSessionDuration,
-		IdleTimeout:        cfg.AgentChatIdleTimeout,
-	}
-}
-
-func (h *Handler) agentChatSnapshotConfig() agentChatSnapshotConfig {
-	return agentChatSnapshotConfigFromServer(h.config.Server)
-}
-
-func renderAgentChatSession(session agentchat.Session, limits agentChatSnapshotConfig) AgentChatSessionItem {
-	messages := make([]AgentChatMessageItem, 0, len(session.Messages))
-	for _, message := range session.Messages {
-		messages = append(messages, AgentChatMessageItem{
-			ID:              message.ID,
-			RunID:           message.RunID,
-			RequestID:       message.RequestID,
-			TraceID:         message.TraceID,
-			SpanID:          message.SpanID,
-			Role:            message.Role,
-			Content:         message.Content,
-			RawOutput:       message.RawOutput,
-			AdapterID:       message.AdapterID,
-			AdapterName:     message.AdapterName,
-			DriverKind:      message.DriverKind,
-			NativeSessionID: message.NativeSessionID,
-			Status:          message.Status,
-			ExitCode:        message.ExitCode,
-			CostMode:        message.CostMode,
-			Workspace:       message.Workspace,
-			DiffStat:        message.DiffStat,
-			Diff:            message.Diff,
-			CreatedAt:       formatOptionalTime(message.CreatedAt),
-			StartedAt:       formatOptionalTime(message.StartedAt),
-			CompletedAt:     formatOptionalTime(message.CompletedAt),
-			DurationMS:      durationMillis(message.StartedAt, message.CompletedAt),
-			Error:           message.Error,
-			Activities:      renderAgentChatActivities(message.Activities),
-			Usage:           renderAgentChatUsage(message.Usage),
-		})
-	}
-	return AgentChatSessionItem{
-		ID:                   session.ID,
-		Title:                session.Title,
-		AdapterID:            session.AdapterID,
-		DriverKind:           session.DriverKind,
-		NativeSessionID:      session.NativeSessionID,
-		Workspace:            session.Workspace,
-		WorkspaceBranch:      session.WorkspaceBranch,
-		Status:               session.Status,
-		TurnsUsed:            session.TurnsUsed,
-		MaxTurnsPerSession:   limits.MaxTurnsPerSession,
-		SessionStartedAt:     formatOptionalTime(session.CreatedAt),
-		MaxSessionDurationMS: limits.MaxSessionDuration.Milliseconds(),
-		IdleTimeoutMS:        limits.IdleTimeout.Milliseconds(),
-		CreatedAt:            formatOptionalTime(session.CreatedAt),
-		UpdatedAt:            formatOptionalTime(session.UpdatedAt),
-		Messages:             messages,
-	}
-}
-
-func agentChatUsageFromResult(usage agentadapters.Usage) agentchat.Usage {
-	return agentchat.Usage{
-		ContextSize:          usage.ContextSize,
-		ContextUsed:          usage.ContextUsed,
-		ReportedCostAmount:   usage.ReportedCostAmount,
-		ReportedCostCurrency: usage.ReportedCostCurrency,
-	}
-}
-
-func renderAgentChatUsage(usage agentchat.Usage) *AgentChatUsageItem {
-	if usage.Empty() {
-		return nil
-	}
-	return &AgentChatUsageItem{
-		ContextSize:          usage.ContextSize,
-		ContextUsed:          usage.ContextUsed,
-		ReportedCostAmount:   usage.ReportedCostAmount,
-		ReportedCostCurrency: usage.ReportedCostCurrency,
-	}
-}
-
-func (h *Handler) startAgentChatTrace(w http.ResponseWriter, r *http.Request) (*profiler.Trace, context.Context) {
-	requestID := RequestIDFromContext(r.Context())
-	if requestID == "" {
-		requestID = newRequestID()
-	}
-	trace := h.tracer.Start(requestID)
-	ctx := telemetry.WithTraceIDs(r.Context(), trace.TraceID, trace.RootSpanID())
-	w.Header().Set("X-Trace-Id", trace.TraceID)
-	w.Header().Set("X-Span-Id", trace.RootSpanID())
-	return trace, ctx
-}
-
-func agentChatTraceAttrs(session agentchat.Session, adapter agentadapters.Adapter, runID, messageID string, attrs map[string]any) map[string]any {
-	out := map[string]any{
-		telemetry.AttrHecateAgentChatSessionID:  session.ID,
-		telemetry.AttrHecateAgentChatMessageID:  messageID,
-		telemetry.AttrHecateRunID:               runID,
-		telemetry.AttrHecateExecutionKind:       "agent_chat",
-		telemetry.AttrHecateAgentAdapterID:      adapter.ID,
-		telemetry.AttrHecateAgentAdapterName:    adapter.Name,
-		telemetry.AttrHecateAgentAdapterCommand: adapter.Command,
-		telemetry.AttrHecateAgentDriverKind:     adapter.Kind,
-		telemetry.AttrHecateWorkspacePath:       session.Workspace,
-		telemetry.AttrHecateResult:              telemetry.ResultSuccess,
-	}
-	if session.NativeSessionID != "" {
-		out[telemetry.AttrHecateAgentNativeSessionID] = session.NativeSessionID
-	}
-	for key, value := range attrs {
-		out[key] = value
-	}
-	return out
-}
-
-func agentChatTerminalEvent(status string) string {
-	switch status {
-	case "cancelled":
-		return telemetry.EventAgentChatRunCancelled
-	case "failed":
-		return telemetry.EventAgentChatRunFailed
-	default:
-		return telemetry.EventAgentChatRunFinished
-	}
-}
-
-func renderAgentChatActivities(items []agentchat.Activity) []AgentChatActivityItem {
-	if len(items) == 0 {
-		return nil
-	}
-	out := make([]AgentChatActivityItem, 0, len(items))
-	for _, item := range items {
-		out = append(out, AgentChatActivityItem{
-			ID:        item.ID,
-			Type:      item.Type,
-			Status:    item.Status,
-			Kind:      item.Kind,
-			Title:     item.Title,
-			Detail:    item.Detail,
-			CreatedAt: formatOptionalTime(item.CreatedAt),
-		})
-	}
-	return out
-}
-
-func newAgentChatActivity(kind, status, title, detail string) agentchat.Activity {
-	return agentchat.Activity{
-		Type:      kind,
-		Status:    status,
-		Title:     title,
-		Detail:    strings.TrimSpace(detail),
-		CreatedAt: time.Now().UTC(),
-	}
-}
-
-func agentChatActivityFromAdapter(activity agentadapters.Activity) agentchat.Activity {
-	return agentchat.Activity{
-		ID:        strings.TrimSpace(activity.ID),
-		Type:      strings.TrimSpace(activity.Type),
-		Status:    strings.TrimSpace(activity.Status),
-		Kind:      strings.TrimSpace(activity.Kind),
-		Title:     strings.TrimSpace(activity.Title),
-		Detail:    strings.TrimSpace(activity.Detail),
-		CreatedAt: time.Now().UTC(),
-	}
-}
-
-func mergeAgentChatActivity(items []agentchat.Activity, next agentchat.Activity) []agentchat.Activity {
-	if next.Type == "" || (next.ID == "" && next.Title == "") {
-		return items
-	}
-	if next.ID != "" {
-		for i := range items {
-			if items[i].ID == next.ID {
-				if next.Status != "" {
-					items[i].Status = next.Status
-				}
-				if next.Kind != "" {
-					items[i].Kind = next.Kind
-				}
-				if next.Title != "" {
-					items[i].Title = next.Title
-				}
-				if next.Detail != "" {
-					items[i].Detail = next.Detail
-				}
-				items[i].CreatedAt = next.CreatedAt
-				return items
-			}
-		}
-	}
-	if next.Title == "" {
-		return items
-	}
-	return append(items, next)
-}
-
-func finalAgentChatActivityTitle(status string) string {
-	switch status {
-	case "completed":
-		return "Final answer"
-	case "failed":
-		return "Failed"
-	case "cancelled":
-		return "Cancelled"
-	default:
-		return status
-	}
-}
-
-func durationMillis(startedAt, completedAt time.Time) int64 {
-	if startedAt.IsZero() || completedAt.IsZero() || completedAt.Before(startedAt) {
-		return 0
-	}
-	return completedAt.Sub(startedAt).Milliseconds()
-}
-
-// sendAgentChatSSE writes one SSE event to the response. The payload
-// is marshalled to JSON; on encode failure we emit an `event: error`
-// frame so subscribers see something rather than silently stalling.
-// Accepts any payload so the same writer can carry session updates
-// (AgentChatSessionResponse) and approval events
-// (AgentChatApprovalRequestedEvent / AgentChatApprovalResolvedEvent).
-func sendAgentChatSSE(w http.ResponseWriter, flusher http.Flusher, event string, payload any) {
-	data, err := json.Marshal(payload)
-	if err != nil {
-		fmt.Fprintf(w, "event: error\ndata: {\"error\":{\"message\":%q}}\n\n", err.Error())
-		flusher.Flush()
-		return
-	}
-	fmt.Fprintf(w, "event: %s\ndata: %s\n\n", event, data)
-	flusher.Flush()
 }
 
 func isTerminalAgentChatStatus(status string) bool {
